Dolby Surround Audio Effects: Setting pre-gain volume level
Added changes required to set pre-gain volume level.

The Dolby Surround (DS) Audio Processing library expects to be informed about
any attenuation that has been applied to the audio signal, which is input to
the audio processing library. The library is informed by way of the 'pre-gain'
parameter. As DS1 Audio Processing is applied globally in the Android audio
chain, the maximum volume among the tracks is calculated and the value for
this volume level is sent as the pre-gain to the DS Audio Effect.

The DS effect is registered as an Android audio effect in the following file:
frameworks/base/media/libeffects/data/audio_effects.conf
(http://review.omapzoom.org/#/c/26856/).

Change-Id: Ica93c84e78e8461f41e22af0a1ae34d3fe8dafa0